pyproject.toml
[tool.poetry]
name = "dashrep"
version = "0.1.0"
description = "sitrep using dash"
authors = ["Steve Harris <doc@steveharris.me>"]

[tool.poetry.dependencies]
python = "^3.9"
plotly = "^5.3.1"
dash = "^2.0.0"
pandas = "*"
numpy = "*"
python-dotenv = "^0.19.1"
dash-bootstrap-components = "^1.0.0"
Flask = "^2.0.1"
pytest = "^6.2.5"
Faker = "^9.5.2"
requests = "^2.26.0"
dash-daq = "^0.5.0"
arrow = "^1.2.1"
SQLAlchemy = "^1.4.28"

[tool.poetry.dev-dependencies]
jupyterlab = "^3.2.1"
python-lsp-server = "^1.3.1"
python-lsp-black = "^1.0.0"
mypy-ls = "^0.5.1"
pyls-isort = "^0.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
